update modify folder dialog
