+++ packages/veritone-react-common/src/components/FolderTree/Modals/EditFolder/index.js
@@ -2,23 +2,40 @@
 import React from 'react';
 import Button from '@material-ui/core/Button';
 import cx from 'classnames';
-import { get, isNil } from 'lodash';
-import { shape, bool, func } from 'prop-types';
+import { get, isNil, omit, isEmpty } from 'lodash';
+import { shape, bool, func, arrayOf, string } from 'prop-types';
 import TextField from '@material-ui/core/TextField';
 import Dialog from '@material-ui/core/Dialog';
 import DialogActions from '@material-ui/core/DialogActions';
 import DialogContent from '@material-ui/core/DialogContent';
 import DialogTitle from '@material-ui/core/DialogTitle';
+import FolderTree from '../../index';
 import styles from './styles.scss';
 
-export default function ModifyFolder({
+const actionType = {
+  1: 'New Folder',
+  2: 'Move Folder',
+  3: 'Edit Folder'
+}
+
+export default function EditFolder({
   open,
+  isNewFolder = false,
+  type = 'New Folder',
+  isEnableEditName = false,
+  isEnableEditFolder = false,
   currentFolder = {},
   handleClose,
-  handleSubmit
+  handleSubmit,
+  foldersData,
+  onExpand,
+  defaultOpening = [],
+  processingFolder = [],
+  handerClickNewFolder
 }) {
   const [folderName, setFolderName] = React.useState('');
   const [error, setError] = React.useState('');
+  const [selectedFolder, setSelectedFolder] = React.useState({});
   React.useEffect(() => {
     if (!isNil(currentFolder)) {
       const folderName = get(currentFolder, 'name');
@@ -45,10 +62,14 @@ export default function ModifyFolder({
     validate(value);
   };
 
+  const onSelectFolder = folders => {
+    setSelectedFolder(folders);
+  };
+
   const onUpdate = () => {
     validate(folderName);
     if (error === '') {
-      handleSubmit(folderName, currentFolder);
+      handleSubmit(selectedFolder, folderName);
     }
   };
 
@@ -59,6 +80,24 @@ export default function ModifyFolder({
     setError('');
   };
 
+  const foldersDataReprocess = () => {
+    return {
+      ...foldersData,
+      allId: [...foldersData.allId.filter(item => item !== currentFolder.id)],
+      byId: omit(foldersData.byId, currentFolder.id)
+    }
+  };
+
+  const handlerNewFolder = () => {
+    handerClickNewFolder(selectedFolder);
+  }
+
+  const getSubmitStatus = () => {
+    if (type === 3) {
+      return folderName === '';
+    }
+    return folderName === '' || isEmpty(selectedFolder);
+  }
 
   return (
     <div>
@@ -69,22 +108,58 @@ export default function ModifyFolder({
         onClose={handleClose}
         aria-labelledby="create-folder"
       >
-        <DialogTitle id="create-folder">Edit Folder</DialogTitle>
+        <DialogTitle id="create-folder">
+          {actionType[type]}
+        </DialogTitle>
         <DialogContent className={cx(styles['dialog-content'])}>
-          <div className={cx(styles['folder-name-field'])}>
-            <TextField
-              autoFocus
-              margin="dense"
-              id="folder-name"
-              label="Folder Name"
-              type="text"
-              error={error.length !== 0}
-              helperText={error}
-              value={folderName}
-              onChange={onChange}
-              fullWidth
-            />
-          </div>
+          {
+            isEnableEditName && (
+              <div className={cx(styles['folder-name-field'])}>
+                <TextField
+                  autoFocus
+                  margin="dense"
+                  id="folder-name"
+                  label="Folder Name"
+                  type="text"
+                  error={error.length !== 0}
+                  helperText={error}
+                  value={folderName}
+                  onChange={onChange}
+                  fullWidth
+                />
+              </div>
+            )
+          }
+          {isEnableEditFolder && (
+            <div className={cx(styles['action-new-field'])}>
+              <div>Choose a Folder</div>
+              {!isNewFolder && (
+                <Button
+                  color="primary"
+                  className={cx(styles['button-styles'])}
+                  onClick={handlerNewFolder}
+                >
+                  NEW FOLDER
+              </Button>
+              )}
+            </div>
+          )}
+          {isEnableEditFolder && (
+            <div className={cx(styles['folder-tree-card'])}>
+              <FolderTree
+                selectable={false}
+                loading={false}
+                foldersData={foldersDataReprocess()}
+                onChange={onSelectFolder}
+                onExpand={onExpand}
+                isEnableShowContent={false}
+                isEnableShowRootFolder
+                selected={selectedFolder}
+                defaultOpening={defaultOpening}
+                processingFolder={processingFolder}
+              />
+            </div>
+          )}
         </DialogContent>
         <DialogActions>
           <Button
@@ -95,7 +170,7 @@ export default function ModifyFolder({
             Cancel
           </Button>
           <Button
-            disabled={folderName === ''}
+            disabled={getSubmitStatus()}
             className={cx(styles['button-styles'])}
             onClick={onUpdate}
             color="primary"
@@ -107,9 +182,18 @@ export default function ModifyFolder({
     </div>
   );
 }
-ModifyFolder.propTypes = {
+EditFolder.propTypes = {
   open: bool,
+  isEnableEditName: bool,
+  isEnableEditFolder: bool,
+  isNewFolder: bool,
+  type: string,
   currentFolder: shape(Object),
   handleClose: func,
   handleSubmit: func,
+  foldersData: shape(Object),
+  onExpand: func,
+  handerClickNewFolder: func,
+  defaultOpening: arrayOf(Object),
+  processingFolder: arrayOf(Object)
 }
