+++ packages/veritone-react-common/src/components/FolderTree/Modals/EditFolder/styles.scss
@@ -13,7 +13,8 @@
 .folder-tree-card {
   overflow: auto !important;
   max-height: 350px;
-  padding: 5px 0;
+  border: 1px solid #e0e0e0;
+  background-color: #fff;
 }
 .button-styles {
   padding: 6px 16px !important;
