--- packages/veritone-react-common/src/components/FolderTree/Modals/ModifyFolder/styles.scss
@@ -1,20 +1,0 @@
-.folder-name-field {
-  padding-bottom: 20px !important;
-}
-.action-new-field {
-  justify-content: space-between;
-  display: flex;
-  align-items: center;
-  padding-bottom: 20px;
-}
-.dialog-content {
-  overflow: hidden !important;
-}
-.folder-tree-card {
-  overflow: auto !important;
-  max-height: 350px;
-  padding: 5px 0;
-}
-.button-styles {
-  padding: 6px 16px !important;
-}
